+++ requirements-dev.txt
@@ -1,5 +1,4 @@
 pytest
-pytest-qt
 pytest-mock
 pytest-env
 pytest-timeout
