+++ requirements.txt
@@ -1,7 +1,3 @@
-PySide6>=6.7
-PySide6-Addons>=6.7
-pyobjc-core>=12.0; sys_platform == "darwin"
-pyobjc-framework-Cocoa>=12.0; sys_platform == "darwin"
 mcp>=1.9
 uv>=0.7.0
 PyYAML>=6.0
